// apps/project-a/src/shared/types/market-content.types.ts
export interface LoginMarketContent {
  title: string;
  description: string;
  usernameLabel: string;
  passwordLabel: string;
  submitButton: string;
  goBackLabel: string;
}

export interface ProductMarketContent {
  title: string;
  backToProducts: string;
  loginPrompt: string;
  addToCart: string;
  inStock: string;
  outOfStock: string;
  currency: string;
  stockLabel: string;
  brandLabel: string;
  ratingLabel: string;
  discountLabel: string;
  extendedInfoTitle: string;
  loginNow: string;
}

export interface ProductsMarketContent {
  title: string;
  description: string;
  backToHome: string;
  viewDetails: string;
  currency: string;
}